requirements.txt
branca==0.6.0
folium==0.14.0
geopandas==0.13.0
Jinja2==3.1.2
numpy<1.24.3
pandas==2.0.2
Pillow<9.5.0
polygeohasher==0.0.5
pyperclip==1.8.2
Shapely==2.0.1
streamlit==1.29.0
streamlit_folium==0.11.1
jsons<1.6.3
zipfile36==0.1.3
pytest-shutil==1.7.0
temp==2020.7.2